+++ validator/validator.go
@@ -48,12 +48,12 @@ func Cron(value interface{}, _ cty.Path) diag.Diagnostics {
 
 func CronLength(value interface{}, _ cty.Path) diag.Diagnostics {
 	var diags diag.Diagnostics
-	cron := regexp.MustCompile("[^\\s]+").FindAllString(value.(string), -1)
+	cron := strings.Split(value.(string), " ")
 
-	if len(cron) > 7 || len(cron) < 6 {
+	if len(cron) < 6 || len(cron) > 7 {
 		diags = append(diags, diag.Diagnostic{
 			Severity: diag.Error,
-			Summary:  "Invalid Cron expression, value should be between 6 and 7 characters long",
+			Summary:  "Invalid Cron expression, value should be between 6 and 7 parts long",
 			Detail:   fmt.Sprintf("%s is not a valid cron", value),
 		})
 	}